task authorization

+++ app/Http/Controllers/TasksController.php
@@ -7,25 +7,35 @@ use Illuminate\Http\Request;
 use App\Http\Requests;
 
 use App\Task;
+use App\User;
+
 
 class TasksController extends Controller
 {
+    
     public function index()
     {
-        if (\Auth::check()) {
-            $tasks = Task::all();
+        $tasks = [];
+        
+        if (\Auth::check())
+        {
+            $tasks = Task::all()->where('user_id', '=', \Auth::user()->id);
             return view('tasks.index', [
-                'tasks' => $tasks,
+                'tasks' => $tasks, 'newtask' => new Task,
             ]);
         } else {
-            return view('auth.login');
+            return redirect('login');
         }
+            
         
     }
     
     public function show($id) 
     {
         $task = Task::find($id);
+        if ($task->user_id != \Auth::user()->id) {
+            return redirect('/');
+        }
         return view('tasks.show', [
             'task' => $task,
         ]);
@@ -41,11 +51,13 @@ class TasksController extends Controller
     
     public function store(Request $request)
     {
+        
         $this->validate($request, [
             'content' => 'required|max:191',    
         ]);
         
         $task = new Task;
+        $task->user_id = \Auth::user()->id;
         $task->content = $request->content;
         $task->save();
         
@@ -55,6 +67,11 @@ class TasksController extends Controller
     public function edit($id)
     {
         $task = Task::find($id);
+        
+        if ($task->user_id != \Auth::user()->id) {
+            return redirect('/');
+        }
+        
         return view('tasks.edit', [
             'task' => $task,
         ]);
@@ -67,6 +84,11 @@ class TasksController extends Controller
         ]);
         
         $task = Task::find($id);
+        
+        if ($task->user_id != \Auth::user()->id) {
+            return redirect('/');
+        }
+        
         $task->content = $request->content;
         $task->save();
         
@@ -76,6 +98,11 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
+        
+        if ($task->user_id != \Auth::user()->id) {
+            return redirect('/');
+        }
+        
         $task->delete();
         
         return redirect('/');
